feat: update financial indicator calculations to handle NaN values and refactor related structures

+++ src/infrastructure/repositories/statement/structures/jquants_api.rs
@@ -16,12 +16,12 @@ impl From<Response<'_>> for model::RowStatement {
         let disclosed_date = NaiveDate::from_json_string_value("DisclosedDate", value).ok();
         let eps = f64::from_json_string_value("EarningsPerShare", value).ok();
         let bps = f64::from_json_string_value("BookValuePerShare", value).ok();
-        let net_sales = u64::from_json_string_value("NetSales", value).ok();
-        let opp = u64::from_json_string_value("OperatingProfit", value).ok();
-        let orp = u64::from_json_string_value("OrdinaryProfit", value).ok();
-        let profit = u64::from_json_string_value("Profit", value).ok();
-        let equity = u64::from_json_string_value("Equity", value).ok();
-        let total_assets = u64::from_json_string_value("TotalAssets", value).ok();
+        let net_sales = f64::from_json_string_value("NetSales", value).ok();
+        let opp = f64::from_json_string_value("OperatingProfit", value).ok();
+        let orp = f64::from_json_string_value("OrdinaryProfit", value).ok();
+        let profit = f64::from_json_string_value("Profit", value).ok();
+        let equity = f64::from_json_string_value("Equity", value).ok();
+        let total_assets = f64::from_json_string_value("TotalAssets", value).ok();
 
         Self {
             code,
